refactor: share with element
reuse share selection component
