+++ ts/Common/ShareSelection.scss
@@ -1,0 +1,33 @@
+.bbb-selection-container {
+  position: relative;
+
+  input {
+    width: 100%;
+  }
+}
+
+.bbb-selection {
+  position: absolute;
+  width: 100%;
+  background-color: #ffffff;
+  border: 1px solid var(--color-border-dark);
+  max-height: 88px;
+  overflow: auto;
+  box-shadow: 0 5px 10px -5px var(--color-box-shadow);
+  z-index: 100;
+
+  li {
+    color: var(--color-text-lighter);
+    padding: 0 1em;
+    line-height: 44px;
+
+    &.suggestion {
+      color: var(--color-main-text);
+      cursor: pointer;
+
+      &:hover {
+        background-color: var(--color-background-hover);
+      }
+    }
+  }
+}
