+++ ts/Manager/ShareWith.scss
@@ -35,39 +35,6 @@
   }
 }
 
-.bbb-selection-container {
-  position: relative;
-}
-
-.bbb-selection {
-  position: absolute;
-  width: 100%;
-  background-color: #ffffff;
-  border: 1px solid var(--color-border-dark);
-  max-height: 88px;
-  overflow: auto;
-  box-shadow: 0 5px 10px -5px var(--color-box-shadow);
-
-  li {
-    color: var(--color-text-lighter);
-    padding: 0 1em;
-    line-height: 44px;
-
-    &.suggestion {
-      color: var(--color-main-text);
-      cursor: pointer;
-
-      &:hover {
-        background-color: var(--color-background-hover);
-      }
-    }
-  }
-}
-
-.bbb-form-shareWith {
-  margin-top: -1.5em;
-}
-
 .bbb-icon-unselected {
   opacity: 0.2 !important;
