--- ts/Restrictions/ShareWith.scss
@@ -1,82 +1,0 @@
-.bbb-shareWith {
-  margin: 1em 0;
-
-  &__item {
-    height: 44px;
-    display: flex;
-    align-items: center;
-
-    &:hover {
-      background-color: var(--color-background-dark);
-    }
-
-    .avatardiv {
-      height: 32px;
-      width: 32px;
-      overflow: hidden;
-
-      .icon-group-white {
-        display: block;
-        height: 100%;
-        width: 100%;
-        background-color: #a9a9a9;
-      }
-    }
-
-    .icon {
-      height: 44px;
-      width: 44px;
-    }
-
-    &__label {
-      padding: 0 1em;
-      flex-grow: 1;
-    }
-  }
-}
-
-.bbb-selection-container {
-  position: relative;
-
-  input {
-    width: 100%;
-  }
-}
-
-.bbb-selection {
-  position: absolute;
-  width: 100%;
-  background-color: #ffffff;
-  border: 1px solid var(--color-border-dark);
-  max-height: 88px;
-  overflow: auto;
-  box-shadow: 0 5px 10px -5px var(--color-box-shadow);
-  z-index: 100;
-
-  li {
-    color: var(--color-text-lighter);
-    padding: 0 1em;
-    line-height: 44px;
-
-    &.suggestion {
-      color: var(--color-main-text);
-      cursor: pointer;
-
-      &:hover {
-        background-color: var(--color-background-hover);
-      }
-    }
-  }
-}
-
-.bbb-form-shareWith {
-  margin-top: -1.5em;
-}
-
-.bbb-icon-unselected {
-  opacity: 0.2 !important;
-
-  &:hover {
-    opacity: 0.5 !important;
-  }
-}
